Tentando coisas

+++ header/Grafo.hpp
@@ -22,7 +22,7 @@ public:
   Vertice& no(int id);                 
   void encerra_viagem(std::vector<int>& solucao, Vertice& no_ref);                                 // c
   std::pair<int, std::vector<int>> dijkstra(int origem, int destino);                      // m
-  std::pair<std::vector<int>, int> geraSolucao(double alpha);
+  std::pair<std::vector<int>, int> geraSolucao(float alpha);
     std::pair<std::vector<int>, int> geraSolucaoZetsubou();
   void encerra_dia(std::vector<int>& solucao, Vertice& no_ref);
   void resolveProblema_do_Hotel(std::vector<int> destinos_obrigatorios);
